% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CollateData.R
\name{apply_filters}
\alias{apply_filters}
\title{Convenience function to apply a list of filters to a SummarizedExperiment object}
\usage{
apply_filters(se, filters)
}
\arguments{
\item{se}{A SummarizedExperiment object created by \code{MakeSE()}}

\item{filters}{A list of filters to apply. Each filter must contain the elements
\code{filterClass}, \code{filterType} and \code{filterVars}. See \code{?runFilter} for details}
}
\value{
A vector of logicals, with \code{TRUE} indicating events to be retained, and
\code{FALSE} for events to be filtered out
}
\description{
See \code{\link[=runFilter]{runFilter()}} for details regarding filters
}
\examples{
# see ?MakeSE on example code of generating this NxtSE object
se = NxtIRF_example_NxtSE()

# Get NxtIRF recommended filters
filters = get_default_filters()

# Apply all recommended filters:
se.filtered = apply_filters(
    se = se, 
    filters = filters
)
}
\seealso{
\code{\link[=get_default_filters]{get_default_filters()}}, \code{\link[=runFilter]{runFilter()}}
}
