% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DE_IR.R
\name{ASE-methods}
\alias{ASE-methods}
\alias{limma_ASE}
\alias{DESeq_ASE}
\alias{DoubleExpSeq_ASE}
\title{Use Limma, DESeq2 or DoubleExpSeq to test for differential ASEs
(Alternative Splice Events)}
\usage{
limma_ASE(
  se,
  test_factor,
  test_nom,
  test_denom,
  batch1 = "",
  batch2 = "",
  filter_antiover = TRUE,
  filter_antinear = FALSE
)

DESeq_ASE(
  se,
  test_factor,
  test_nom,
  test_denom,
  batch1 = "",
  batch2 = "",
  n_threads = 1,
  filter_antiover = TRUE,
  filter_antinear = FALSE
)

DoubleExpSeq_ASE(
  se,
  test_factor,
  test_nom,
  test_denom,
  filter_antiover = TRUE,
  filter_antinear = FALSE
)
}
\arguments{
\item{se}{The SummarizedExperiment object created by \code{MakeSE()}. To reduce
runtime and false negatives due to multiple testing issues, please filter
the object using \code{apply_filter()}}

\item{test_factor}{A string for the column name which contains the
contrasting variable}

\item{test_nom}{The condition in which to test for differential ASE. Usually
the "treatment" condition}

\item{test_denom}{The condition in which to test against for differential
ASE. Usually the "control" condition}

\item{batch1, batch2}{(Limma and DESeq2 only) One or two columns containing
batch information to normalise against (can be omitted).}

\item{filter_antiover}{Whether to filter out IR events that overlap
antisense genes (for unstranded RNAseq protocols)}

\item{filter_antinear}{Whether to filter out IR events near but not
overlapping antisense genes (for unstranded RNAseq protocols)}

\item{n_threads}{(DESeq_ASE only) How many threads to use for DESeq2
based analysis.}
}
\value{
A data table containing the following:
\itemize{
\item EventName: The name of the ASE event
\item EventType: The type of event. IR = intron retention, MXE = mutually
exclusive event, SE = skipped exon, AFE = alternate first exon, ALE =
alternate last exon, A5SS / A3SS = alternate 5' / 3' splice site
\item EventRegion: The genomic coordinates the event occupies.
\item NMD_direction: Indicates whether one isoform is a NMD substrate. +1 means
included isoform is NMD, -1 means the excluded isoform is NMD, and 0
means neither (or both) are NMD substrates
\item AvgPSI_nom, Avg_PSI_denom: the average percent spliced in / percent
IR levels for the two conditions being contrasted. \code{nom} and \code{denom} in
column names are replaced with the condition names
}

\strong{limma specific output}
\itemize{
\item logFC, AveExpr, t, P.Value, adj.P.Val, B: limma topTable columns of
differential ASE. See \code{?limma::topTable} for details.
\item inc/exc_(logFC, AveExpr, t, P.Value, adj.P.Val, B): limma results
for differential testing for raw included / excluded counts only
}

\strong{DESeq2 specific output}
\itemize{
\item baseMean, log2FoldChange, lfcSE, stat, pvalue, padj:
DESeq2 results columns for differential ASE; see \code{?DESeq2::results} for
details.
\item inc/exc_(baseMean, log2FoldChange, lfcSE, stat, pvalue, padj):
DESeq2 results for differential testing for
raw included / excluded counts only
}

\strong{DoubleExp specific output}
\itemize{
\item MLE_nom, MLE_denom: Expectation values for the two groups. \code{nom} and
\code{denom} in column names are replaced with the condition names
\item MLE_LFC: Log2-fold change of the MLE
\item P.Value, adj.P.Val: Nominal and BH-adjusted P values
\item n_eff: Number of effective samples (i.e. non-zero or non-unity PSI)
\item mDepth: Mean Depth of splice coverage in each of the two groups.
\item Dispersion_Reduced, Dispersion_Full: Dispersion values for reduced and
full models. See \code{?DoubleExpSeq::DBGLM1} for details.
}
}
\description{
Use Limma, DESeq2 or DoubleExpSeq to test for differential ASEs
(Alternative Splice Events)
}
\section{Functions}{
\itemize{
\item \code{limma_ASE}: Use limma to perform differential ASE analysis of
a filtered NxtSE object

\item \code{DESeq_ASE}: Use DESeq2 to perform differential ASE analysis of
a filtered NxtSE object

\item \code{DoubleExpSeq_ASE}: Use DoubleExpSeq to perform differential ASE analysis
of a filtered NxtSE object (uses double exponential beta-binomial model)
to estimate group dispersions, followed by LRT
}}

\examples{
# see ?MakeSE on example code of generating this NxtSE object
se = NxtIRF_example_NxtSE()

colData(se)$treatment = rep(c("A", "B"), each = 3)

if("limma" \%in\% rownames(installed.packages())) {
    res_limma = limma_ASE(se, "treatment", "A", "B")
}

if("DESeq2" \%in\% rownames(installed.packages())) {
    res_DESeq = DESeq_ASE(se, "treatment", "A", "B")
}

if("DoubleExpSeq" \%in\% rownames(installed.packages())) {
    res_DES = DoubleExpSeq_ASE(se, "treatment", "A", "B")
}
}
