% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CollateData.R
\name{Find_IRFinder_Output}
\alias{Find_IRFinder_Output}
\title{Convenience Function to find IRFinder output and COV files}
\usage{
Find_IRFinder_Output(sample_path, ...)
}
\arguments{
\item{sample_path}{The path in which to recursively search for BAM files}

\item{...}{Additional parameters to pass into \code{FindSamples()}}
}
\value{
A 3-column data frame with the first column containing
the sample name, and the second column being the IRFinder main output path,
and the third column being the COV file path.
}
\description{
Runs FindSamples to find IRFinder .txt.gz and .cov files.
Assumes file names are names of samples.
}
\examples{
expr = Find_IRFinder_Output(file.path(tempdir(), "IRFinder_output"))
}
\seealso{
\link{FindSamples}
}
