% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CollateData.R
\name{FindSamples}
\alias{FindSamples}
\title{Searches a specified path for files of a particular file pattern}
\usage{
FindSamples(
  sample_path,
  suffix = ".txt.gz",
  suffix_type = "path",
  use_subdir = FALSE
)
}
\arguments{
\item{sample_path}{The path in which to recursively search for files
that match the given \code{suffix}}

\item{suffix}{A vector of or or more strings that specifies the file suffix
(e.g. '.bam' denotes BAM files, whereas ".txt.gz" denotes gzipped txt files).}

\item{suffix_type}{A vector of string that determines the column
names of the files retrieved by \code{suffix}. Must be the same length as \code{suffix}}

\item{use_subdir}{Whether to assume the directory name containing
the found files denote the sample name. If \code{FALSE}, the base name
of the file is assumed to be the sample name. See below example.}
}
\value{
A 2-column data frame with the first column containing
the sample name, and the second column being the file path.
}
\description{
This convenience function identifies files with the specified suffix.
The sample name is assumed to be the name of the file minus its suffix
(\code{use_subdir = FALSE}), but can also be changed to be the names
of the parent directory (\code{use_subdir = TRUE}). See example below.
}
\examples{
# Return all BAM files using file names as sample names
bam_path = dirname(example_bams())[1]
df = FindSamples(sample_path = bam_path, 
  suffix = ".bam", suffix_type = "bam_file", use_subdir = FALSE)
}
