% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildRef.R
\name{GenerateMappabilityBED}
\alias{GenerateMappabilityBED}
\title{Generate a BED file containing low mappability regions from an aligned genome}
\usage{
GenerateMappabilityBED(BAM = "", out.bed, threshold = 4)
}
\arguments{
\item{BAM}{The file name of the aligned BAM file}

\item{out.bed}{The output file name}

\item{threshold}{Regions with this read depth (or below) are defined as low
mappability regions}
}
\value{
None. This function writes a 3-column BED file to \code{out.bed}
}
\description{
NxtIRF uses IRFinder's method of empirically defining low-mappability
regions. This is done by aligning artificial reads derived from a given
genome (alternate +/- strands, with set length and stride), back onto
the genome with the aligner. Genomic regions where the derived read cannot be
mapped back to itself are low-mappability regions
}
\examples{
# GenerateMappabilityBED(BAM = "aligned.BAM",
#   out.bed = file.path(tempdir(), "Reference", "Mappability.bed")
# )
}
