% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildRef.R
\name{Get_Genome}
\alias{Get_Genome}
\title{Gets a genome object from NxtIRF reference.}
\usage{
Get_Genome(reference_path)
}
\arguments{
\item{reference_path}{The directory to the NxtIRF reference}
}
\value{
A TwoBitFile object containing the genome sequence
}
\description{
Gets a genome object from NxtIRF reference.
}
\examples{
genome = Get_Genome(file.path(tempdir(), "Reference"))
}
