% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{GetCoverage}
\alias{GetCoverage}
\title{Calls NxtIRF's C++ function to retrieve coverage}
\usage{
GetCoverage(file, seqname = "", start = 0, end = 0, strand = c("*", "+", "-"))
}
\arguments{
\item{file}{The file name of the COV file}

\item{seqname}{Either blank, or a character string denoting the chromosome 
name}

\item{start}{The 0-based start coordinate}

\item{end}{The 0-based end coordinate}

\item{strand}{Either "*", "+", or "-"}
}
\value{
If seqname is left as "", returns an RLEList of the whole BAM file.
  If seqname and coordinates are given, returns an RLE containing the
  chromosome coordinate. Coordinates outside the given range will be set to 0
}
\description{
This function returns an RLE or RLEList containing coverage data from the
given COV file
}
\examples{
se <- NxtIRF_example_NxtSE()

cov_file <- covfile(se)[1]

cov <- GetCoverage(cov_file, seqname = "chrZ", 
  start = 10000, end = 20000,
  strand = "*"
)
}
