% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{IsCOV}
\alias{IsCOV}
\title{Validates the given file as a valid COV file}
\usage{
IsCOV(coverage_files)
}
\arguments{
\item{coverage_files}{A vector containing the file names of files to be
checked}
}
\value{
\code{TRUE} if all files are valid COV files. \code{FALSE} otherwise
}
\description{
Validates the given file as a valid COV file
}
\examples{
se = NxtIRF_example_NxtSE()

cov_files = covfile(se)

IsCOV(cov_files) # returns true if these are true COV files
}
