% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DE_IR.R
\name{limma_ASE}
\alias{limma_ASE}
\title{Use Limma to test for differential ASE (Alternative Splice Event)}
\usage{
limma_ASE(
  se,
  test_factor,
  test_nom,
  test_denom,
  batch1 = "",
  batch2 = "",
  filter_antiover = TRUE,
  filter_antinear = FALSE,
  filter_annotated_IR = FALSE
)
}
\arguments{
\item{se}{The SummarizedExperiment object created by `MakeSE()`. To reduce runtime and false negatives
due to multiple testing issues, please filter the object using `apply_filter()`}

\item{test_factor}{A string for the column name which contains the contrasting variable}

\item{test_nom}{The condition in which to test for differential ASE. Usually the "treatment" condition}

\item{test_denom}{The condition in which to test against for differential ASE. Usually the "control" condition}

\item{batch1, batch2}{One or two columns containing batch information to normalise against (can be omitted)}

\item{filter_antiover}{Whether to filter out IR events that overlap antisense genes (for unstranded RNAseq protocols)}

\item{filter_antinear}{Whether to filter out IR events near but not overlapping antisense genes 
(for unstranded RNAseq protocols)}

\item{filter_annotated_IR}{Whether to filter out IR events that are already annotated exons
(after doing so, all IR events will be unannotated - i.e. constitutionally spliced introns))}
}
\value{
A data table containing the following:
  EventName: The name of the ASE event\cr\cr
  EventType: The type of event. IR = intron retention, MXE = mutually exclusive event, SE = skipped exon,
    AFE = alternate first exon, ALE = alternate last exon, A5SS / A3SS = alternate 5' / 3' splice site\cr\cr
  EventRegion: The genomic coordinates the event occupies.\cr\cr  
  NMD_direction: Indicates whether one isoform is a NMD substrate. +1 means included isoform is NMD, 
    -1 means the excluded isoform is NMD, and 0 means neither (or both) are NMD substrates\cr\cr
  AvgPSI_nom, Avg_PSI_denom: the average percent spliced in / percent intron retention levels for the
  two conditions being contrasted\cr\cr
  logFC, AveExpr, t, P.Value, adj.P.Val, B: limma topTable columns of limma results. See `?limma::topTable`\cr\cr
  inc/exc_(logFC, AveExpr, t, P.Value, adj.P.Val, B): limma results for differential testing for
    raw included / excluded counts only\cr\cr
}
\description{
Use Limma to test for differential ASE (Alternative Splice Event)
}
\examples{
# see ?MakeSE on example code of generating this NxtSE object
se = NxtIRF_example_NxtSE()

colData(se)$treatment = rep(c("A", "B"), each = 3)

res = limma_ASE(se, "treatment", "A", "B")
}
