% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dash_utils.R
\name{make_diagonal}
\alias{make_diagonal}
\title{Constructs a data frame containing average PSI values of the two contrasted conditions}
\usage{
make_diagonal(
  se,
  event_list,
  condition,
  nom_DE,
  denom_DE,
  depth_threshold = 10,
  logit_max = 5
)
}
\arguments{
\item{se}{A NxtIRF SummarizedExperiment}

\item{event_list}{A character vector containing the row names of ASE events (as given by the
\code{EventName} column of differential ASE results table using \code{limma_ASE()} or \code{DESeq_ASE()})}

\item{condition}{The name of the column containing the condition values in \code{SummarizedExperiment::colData(se)}}

\item{nom_DE}{The condition to be contrasted, e.g. \code{nom_DE = "treatment"}}

\item{denom_DE}{The condition to be contrasted against, e.g. \code{denom_DE = "control"}}

\item{depth_threshold}{(default = 10) Samples with the number of reads supporting either included or excluded
isoforms below this values are excluded}

\item{logit_max}{(default = 5). This function works by converting all PSI values as logit(PSI), taking the average,
then converting back to PSI using inverse logit. Because logit(0) == -Inf and logit(1) = Inf,
this function caps logit values using logit_max}
}
\value{
A a 3 column data frame, with the first column containing \code{event_list} list of ASE events,
and the last 2 columns containing the average PSI values of the nominator and denominator conditions
}
\description{
This function takes an input SummarizedExperiment \code{se}, a list of alternative
splicing events \code{event_list}, the \code{condition} column containing the experimental annotation,
and the nominator \code{nom_DE} and denominator \code{denom_DE}
It returns a 3 column data frame, with the first column containing \code{event_list} list of ASE events,
and the last 2 columns containing the average PSI values of the nominator and denominator conditions
Note that this function takes the geometric mean of PSI, by first converting all values to logit(PSI),
taking the average logit(PSI) values of each condition, and then converting back to PSI using inv.logit.
}
\examples{
se = NxtIRF_example_NxtSE()

event_list = rowData(se)$EventName

colData(se)$treatment = rep(c("A", "B"), each = 3)

diag_values <- make_diagonal(se, event_list,
  condition = "treatment", nom_DE = "A", denom_DE = "B"
)
}
