% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{NxtFilter-class}
\alias{NxtFilter-class}
\title{The NxtFilterVars class}
\description{
This is a simple class to contain a number of filter settings
for the filtering of alternative splicing and intron retention
}
