% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globals.R
\name{NxtIRF.CoordToGR}
\alias{NxtIRF.CoordToGR}
\title{Converts an IGV-style coordinate to a GenomicRanges object}
\usage{
NxtIRF.CoordToGR(coordinates)
}
\arguments{
\item{coordinates}{A vector of strings containing the coordinates
to be converted}
}
\value{
A GRanges object that corresponds to the given coordinates
}
\description{
IGV-style coordinates typically have the syntax
\code{seqnames:start-end/strand}\cr\cr
For example: "chr3:10550-10730/+" or "X:51231-51330/-"
}
\examples{
se = NxtIRF_example_NxtSE()

coordinates = rowData(se)$EventRegion

coords.gr = NxtIRF.CoordToGR(coordinates)
}
