% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dash.R
\name{nxtIRF}
\alias{nxtIRF}
\title{Launches the NxtIRF Graphics User Interface using Shiny Dashboard}
\usage{
nxtIRF()
}
\value{
None
}
\description{
This function launches the NxtIRF interactive app using Shiny Dashboard
}
\examples{
# nxtIRF() # Launches interactive ShinyDashboard NxtIRF app
}
