% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globals.R
\name{NxtIRF_example_bams}
\alias{NxtIRF_example_bams}
\title{Retrieves a data frame containing names and paths of example BAM files}
\usage{
NxtIRF_example_bams()
}
\value{
A data frame, with the columns being the sample names and paths
  of the example BAM dataset aligned to the mock chromosome
}
\description{
Intended to be run using the NxtIRF mock reference genome and gene 
annotations
}
\examples{
bams = NxtIRF_example_bams()
}
