% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globals.R
\name{NxtIRF_example_NxtSE}
\alias{NxtIRF_example_NxtSE}
\title{Retrieves a NxtSE object created from NxtIRF's example bam and mock reference}
\usage{
NxtIRF_example_NxtSE()
}
\value{
A NxtSE object returned after creating and using the mock reference
  to run IRFinder, CollateData and MakeSE on the example bam files
}
\description{
Retrieves a NxtSE object created from NxtIRF's example bam and mock reference
}
\examples{
se = NxtIRF_example_NxtSE()
}
\seealso{
[NxtIRF_example_bams()], [mock_genome()], [mock_gtf()]

[BuildReference()], [IRFinder()], [CollateData()], [MakeSE()]
}
