% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{NxtSE-class}
\alias{NxtSE-class}
\title{The NxtSE class methods}
\description{
The NxtSE class inherits from the \linkS4class{SummarizedExperiment} 
class and is defined to represent that it is constructed from `MakeSE()`.
}
