% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dash_utils.R
\name{Plot_Junctions}
\alias{Plot_Junctions}
\title{Draws an arc plot representing the spliced reads of a given sample}
\usage{
Plot_Junctions(fst_path, seqnames, start, end, strand, sample_name)
}
\arguments{
\item{fst_path}{The path to the output generated by [CollateData()]}

\item{seqnames, start, end, strand}{The region to display. `seqnames` and
`strand` should be characters, whereas start and end should be integers}

\item{sample_name}{The sample name.}
}
\value{
A ggplot object containing the rendered arc plot.
}
\description{
NB Experimental / WIP
}
