% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dash_utils.R
\name{prepare_covplot_data}
\alias{prepare_covplot_data}
\title{Generates the requisite data for Plot_Coverage()}
\usage{
prepare_covplot_data(reference_path)
}
\arguments{
\item{reference_path}{The directory containing the NxtIRF reference}
}
\value{
A list, containing the genome, gene_list, elem.DT (containing information about
  introns, exons and UTRs, and transcripts.DT which contains a list of transcripts
}
\description{
Generates the requisite data for Plot_Coverage()
}
