% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STAR_utils.R
\name{STAR-methods}
\alias{STAR-methods}
\alias{STAR_buildRef}
\alias{STAR_align_experiment}
\alias{STAR_align_fastq}
\title{STAR wrapper for building reference for STAR, and aligning raw 
sequence (FASTQ or FASTA) files.}
\usage{
STAR_buildRef(
  reference_path,
  STAR_ref_path = file.path(reference_path, "STAR"),
  also_generate_mappability = TRUE,
  mappability_reads_fasta = file.path(reference_path, "Mappability", "Reads.fa"),
  mappability_depth_threshold = 4,
  sjdbOverhang = 150,
  n_threads = 4,
  ...
)

STAR_align_experiment(
  Experiment,
  STAR_ref_path,
  BAM_output_path,
  trim_adaptor = "AGATCGGAAG",
  two_pass = FALSE,
  n_threads = 4
)

STAR_align_fastq(
  STAR_ref_path,
  BAM_output_path,
  fastq_1 = c("./sample_1.fq"),
  fastq_2 = NULL,
  two_pass = FALSE,
  trim_adaptor = "AGATCGGAAG",
  memory_mode = "NoSharedMemory",
  additional_args = NULL,
  n_threads = 4
)
}
\arguments{
\item{reference_path}{The path to the reference. 
\code{GetReferenceResource()} must first be run using this path
as its \code{reference_path}}

\item{STAR_ref_path}{(Default - the "STAR" subdirectory under 
\code{reference_path}) The directory containing the STAR reference to be
used or to contain the newly-generated STAR reference}

\item{also_generate_mappability}{Whether \code{STAR_buildRef()} also
calculates Mappability Exclusion regions.}

\item{mappability_reads_fasta}{The path to the FASTA file containing reads
generated by \code{GenerateMappabilityReads()}. 
The default value is the
default path for mappability reads generated by 
\code{GetReferenceResource()} when \code{generate_mappability_reads}
is set to \code{TRUE}}

\item{mappability_depth_threshold}{(Default 4) The depth of mapped reads
threshold at or below which Mappability exclusion regions are defined. See
\code{GenerateMappabilityReads()}.}

\item{sjdbOverhang}{A STAR setting indicating the length of the donor / 
acceptor sequence on each side of the junctions. Ideally equal to 
(mate_length - 1). See the STAR aligner manual for details.}

\item{n_threads}{The number of threads to run the STAR aligner.}

\item{Experiment}{A two or three-column data frame with the columns denoting
sample names, forward-FASTQ and reverse-FASTQ files. This can be 
conveniently generated using \code{Find_FASTQ()}}

\item{BAM_output_path}{The path under which STAR outputs the aligned BAM
files. In \code{STAR_align_experiment()}, STAR will output aligned
BAMS inside subdirectories of this folder, named by sample names. In
\code{STAR_align_fastq()}, STAR will output directly into this path.}

\item{trim_adaptor}{The sequence of the Illumina adaptor to trim via STAR's
\code{--clip3pAdapterSeq} option}

\item{two_pass}{Whether to use two-pass mapping. In 
\code{STAR_align_experiment()}, STAR will first align every sample
and generate a list of splice junctions but not BAM files. The junctions
are then given to STAR to generate a temporary genome (contained within
\code{_STARgenome}) subdirectory within that of the first sample), using
these junctions to improve novel junction detection. In 
\code{STAR_align_fastq()}, STAR will run \code{--twopassMode Basic}}

\item{fastq_1, fastq_2}{Character vectors giving the path(s) of one or more
FASTQ (or FASTA) files to be aligned. If single reads are to be aligned,
omit \code{fastq_2}}

\item{memory_mode}{The parameter to be parsed to \code{--genomeLoad}; either
\code{NoSharedMemory} or \code{LoadAndKeep} are used.}

\item{additional_args}{A character vector of additional arguments to be
parsed into STAR.}
}
\value{
None. STAR will output files into the given output directories.
}
\description{
These functions run the STAR aligner to build a STAR genome reference,
(optionally) calculate mappability exclusion regions, and align one or more
FASTQ files (single or paired) to the generated genome. These functions only
work on Linux-based systems with STAR installed (and accessible via $PATH)
}
\section{Functions}{
\itemize{
\item \code{STAR_buildRef}: Creates a STAR genome reference

\item \code{STAR_align_experiment}: Aligns multiple sets of FASTQ files, separated by
sample names

\item \code{STAR_align_fastq}: Aligns a single sample (with single or paired FASTQ
or FASTA files)
}}

\seealso{
\link{BuildReference} for \code{GetReferenceResource()}\cr\cr
\link{Find_Samples} for \code{Find_FASTQ()}\cr\cr
\link{Mappability-methods} for \code{GenerateMappabilityReads()}\cr\cr
}
