% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DE_IR.R
\name{ASE-methods}
\alias{ASE-methods}
\alias{limma_ASE}
\alias{DESeq_ASE}
\alias{DoubleExpSeq_ASE}
\title{Use Limma, DESeq2 or DoubleExpSeq to test for differential ASEs
(Alternative Splice Events)}
\usage{
limma_ASE(
  se,
  test_factor,
  test_nom,
  test_denom,
  batch1 = "",
  batch2 = "",
  filter_antiover = TRUE,
  filter_antinear = FALSE,
  filter_annotated_IR = FALSE
)

DESeq_ASE(
  se,
  test_factor,
  test_nom,
  test_denom,
  batch1 = "",
  batch2 = "",
  n_threads = 1,
  filter_antiover = TRUE,
  filter_antinear = FALSE,
  filter_annotated_IR = FALSE
)

DoubleExpSeq_ASE(
  se,
  test_factor,
  test_nom,
  test_denom,
  filter_antiover = TRUE,
  filter_antinear = FALSE,
  filter_annotated_IR = FALSE
)
}
\arguments{
\item{se}{The SummarizedExperiment object created by \code{MakeSE()}. To reduce
runtime and false negatives due to multiple testing issues, please filter
the object using \code{apply_filter()}}

\item{test_factor}{A string for the column name which contains the
contrasting variable}

\item{test_nom}{The condition in which to test for differential ASE. Usually
the "treatment" condition}

\item{test_denom}{The condition in which to test against for differential
ASE. Usually the "control" condition}

\item{batch1, batch2}{(Limma and DESeq2 only) One or two columns containing
batch information to normalise against (can be omitted).}

\item{filter_antiover}{Whether to filter out IR events that overlap
antisense genes (for unstranded RNAseq protocols)}

\item{filter_antinear}{Whether to filter out IR events near but not
overlapping antisense genes (for unstranded RNAseq protocols)}

\item{filter_annotated_IR}{Whether to filter out IR events that are already
annotated exons (after doing so, all IR events will be unannotated -
i.e. constitutionally spliced introns))}

\item{n_threads}{(DESeq_ASE only) How many threads to use for DESeq2
based analysis.}
}
\value{
A data table containing the following:
EventName: The name of the ASE event\cr\cr
EventType: The type of event. IR = intron retention, MXE = mutually
exclusive event, SE = skipped exon, AFE = alternate first exon, ALE =
alternate last exon, A5SS / A3SS = alternate 5' / 3' splice site\cr\cr
EventRegion: The genomic coordinates the event occupies.\cr\cr
NMD_direction: Indicates whether one isoform is a NMD substrate. +1 means
included isoform is NMD, -1 means the excluded isoform is NMD, and 0
means neither (or both) are NMD substrates\cr\cr
AvgPSI_nom, Avg_PSI_denom: the average percent spliced in / percent intron
retention levels for the two conditions being contrasted\cr\cr
(LIMMA SPECIFIC OUTPUT)\cr\cr
logFC, AveExpr, t, P.Value, adj.P.Val, B: limma topTable columns of
limma results. See \code{?limma::topTable}\cr\cr
inc/exc_(logFC, AveExpr, t, P.Value, adj.P.Val, B): limma results
for differential testing for raw included / excluded counts only\cr\cr
(DESEQ2 SPECIFIC OUTPUT)\cr\cr
baseMean, log2FoldChange, lfcSE, stat, pvalue, padj:
DESeq2 results columns See \code{?DESeq2::results}\cr\cr
inc/exc_(baseMean, log2FoldChange, lfcSE, stat, pvalue, padj):
DESeq2 results for differential testing for
raw included / excluded counts only\cr\cr
(DoubleExp SPECIFIC OUTPUT)\cr\cr
MLE: Expectation values for the two groups\cr\cr
MLE_LFC: Log2-fold change of the MLE\cr\cr
P.Value, adj.P.Val: Nominal and BH-adjusted P values\cr\cr
n_eff: Number of effective samples (i.e. non-zero or non-unity PSI)\cr\cr
mDepth: Mean Depth of splice coverage\cr\cr
Dispersion_Reduced, Dispersion_Full: Dispersion values for reduced and
full models
}
\description{
Use Limma, DESeq2 or DoubleExpSeq to test for differential ASEs
(Alternative Splice Events)
}
\section{Functions}{
\itemize{
\item \code{limma_ASE}: Use limma to perform differential ASE analysis of
a filtered NxtSE object

\item \code{DESeq_ASE}: Use DESeq2 to perform differential ASE analysis of
a filtered NxtSE object

\item \code{DoubleExpSeq_ASE}: Use DoubleExpSeq to perform differential ASE analysis 
of a filtered NxtSE object (uses double exponential beta-binomial model)
to estimate group dispersions, followed by LRT
}}

\examples{
# see ?MakeSE on example code of generating this NxtSE object
se = NxtIRF_example_NxtSE()

colData(se)$treatment = rep(c("A", "B"), each = 3)

if("limma" \%in\% rownames(installed.packages())) {
    res_limma = limma_ASE(se, "treatment", "A", "B")
}

if("DESeq2" \%in\% rownames(installed.packages())) {
    res_DESeq = DESeq_ASE(se, "treatment", "A", "B")
}

if("DoubleExpSeq" \%in\% rownames(installed.packages())) {
    res_DES = DoubleExpSeq_ASE(se, "treatment", "A", "B")
}
}
