% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{example-NxtIRF-data}
\alias{example-NxtIRF-data}
\alias{NxtIRF_example_bams}
\alias{NxtIRF_example_NxtSE}
\title{NxtIRF Example BAMs and NxtSE Experiment Object}
\usage{
NxtIRF_example_bams()

NxtIRF_example_NxtSE()
}
\value{
See \code{NxtIRF_example_bams()} returns a 2-column data frame containing
sample names and BAM paths of the example dataset. \code{NxtIRF_example_NxtSE()}
returns a NxtSE object.
}
\description{
6 example bam files based on samples from the
Leucegene dataset (GSE67039). Bam files are constructed
based on the complete bam files of 6 samples from Leucegene,
subsetted by regions containing the 7 above genes. Then, the reads of these
subsetted BAMs were realigned to the NxtIRF example reference using
STAR.\cr\cr
A NxtSE SummarizedExperiment object was constructed by running \code{IRFinder()},
\code{CollateData()} and \code{MakeSE()}
on the example BAM files, using the NxtIRF example reference. A full pipeline
of how this NxtSE was generated can be found in
\code{system.file("scripts", "make_data.R", package = "NxtIRFcore")}
}
\section{Functions}{
\itemize{
\item \code{NxtIRF_example_bams}: Returns a 2-column data frame, containing 
sample names and sample paths (in tempdir()) of example BAM files

\item \code{NxtIRF_example_NxtSE}: Returns a (in-memory / realized) NxtSE object 
generated using the NxtIRF mock reference and example BAM files
}}

\examples{

# returns a data frame with the first column as sample names, and the 
# second column as BAM paths

NxtIRF_example_bams() 

# Returns a NxtSE object created by the example bams aligned to the
# mock NxtSE reference

se = NxtIRF_example_NxtSE() 

}
\references{
Generation of the mappability files was performed using NxtIRF using
a method analogous to that described in:

Middleton R, Gao D, Thomas A, Singh B, Au A, Wong JJ, Bomane A, Cosson B,
Eyras E, Rasko JE, Ritchie W.
IRFinder: assessing the impact of intron retention on mammalian gene
expression.
Genome Biol. 2017 Mar 15;18(1):51.
\doi{10.1186/s13059-017-1184-4}
}
\seealso{
\code{\link[=BuildReference]{BuildReference()}}, \code{\link[=IRFinder]{IRFinder()}}, \code{\link[=CollateData]{CollateData()}}, \code{\link[=MakeSE]{MakeSE()}}
}
\keyword{package}
