% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/File_finders.R
\name{Find_Samples}
\alias{Find_Samples}
\alias{Find_FASTQ}
\alias{Find_Bams}
\alias{Find_IRFinder_Output}
\title{Convenience Function to (recursively) find all files in a folder.}
\usage{
Find_Samples(sample_path, suffix = ".txt.gz", level = 0)

Find_FASTQ(
  sample_path,
  paired = TRUE,
  fastq_suffix = c("fastq", "fq", "fastq.gz", "fq.gz"),
  level = 0
)

Find_Bams(sample_path, level = 0)

Find_IRFinder_Output(sample_path, level = 0)
}
\arguments{
\item{sample_path}{The path in which to recursively search for files
that match the given \code{suffix}}

\item{suffix}{A vector of or or more strings that specifies the file suffix
(e.g. 'bam' denotes BAM files, whereas "txt.gz" denotes gzipped txt
files).}

\item{level}{Whether sample names can be found in the file names themselves
(level = 0), or their parent directory (level = 1). Potentially parent
of parent directory (level = 2). Support max level <= 3 (for sanity).}

\item{paired}{Whether to expect single FASTQ files (of the format
"sample.fastq"), or
paired files (of the format "sample_1.fastq", "sample_2.fastq")}

\item{fastq_suffix}{The name of the FASTQ suffix. Options are:
"fastq", "fastq.gz", "fq", or "fq.gz"}
}
\value{
A multi-column data frame with the first column containing
the sample name, and subsequent columns being the file paths with suffix
as determined by \code{suffix}.
}
\description{
Often, output files (whether it be raw sequencing files, aligned sequences)
in BAM files, or IRFinder output files, are stored in a single folder.
Sometimes, these are named according to the samples they represent. Other
times, they have generic names but are partitioned in sub-folders named
by sample names. This function (recursively) finds all files and extracts
sample names assuming the files are named by sample names
(\code{level = 0}). Alternately, the names can be derived from the
subfolders one level higher (\code{level = 1})
}
\section{Functions}{
\itemize{
\item \code{Find_FASTQ}: Returns all FASTQ files in a given folder

\item \code{Find_Bams}: Returns all BAM files in a given folder

\item \code{Find_IRFinder_Output}: Returns all IRFinder output files in a given folder
}}

\examples{
# Retrieve all BAM files in a given folder, named by sample names
bam_path = tempdir()
example_bams(path = bam_path)
df.bams = Find_Samples(sample_path = bam_path, 
  suffix = ".bam", level = 0)
# equivalent to:
df.bams = Find_Bams(bam_path, level = 0)

# Retrieve all IRFinder output files in a given folder, 
# named by sample names

expr = Find_IRFinder_Output(file.path(tempdir(), "IRFinder_output"))
}
