% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CollateData.R
\name{IRFinder}
\alias{IRFinder}
\title{A wrapper function to call NxtIRF/IRFinder}
\usage{
IRFinder(
  bamfiles = "Unsorted.bam",
  sample_names = "sample1",
  reference_path = "./Reference",
  output_path = "./IRFinder_Output",
  n_threads = 1,
  run_featureCounts = FALSE
)
}
\arguments{
\item{bamfiles}{The file names of 1 or more BAM files}

\item{sample_names}{The sample names of the given BAM files. Must
be a vector of the same length as \code{bamfiles}}

\item{reference_path}{The directory of the NxtIRF reference}

\item{output_path}{The directory where NxtIRF/IRFinder output
should be stored}

\item{n_threads}{The number of threads to use. On Linux / Windows, this will
use OpenMP from within the C++ subroutine. On Macs, BiocParallel
MulticoreParam will be used on single-threaded NxtIRF/IRFinder}

\item{run_featureCounts}{Whether this function will run
\code{Rsubread::featureCounts()} on the BAM files. If so, the output will be
saved to "main.FC.Rds" in the output directory as a list object}
}
\value{
None. \code{IRFinder()} will save output to \code{output_path}. \cr\cr
sample.txt.gz: The main IRFinder output file containing the quantitation
of IR and splice junctions, as well as QC information\cr\cr
sample.cov: Contains coverage information in compressed binary. This
format is 5-10X faster than BigWig format (see \code{\link[=GetCoverage]{GetCoverage()}})\cr\cr
main.FC.Rds: A single file containing gene counts for the whole dataset
(only if \code{run_featureCounts == TRUE})
}
\description{
This function calls IRFinder on one or more BAM files.
}
\examples{
bams = NxtIRF_example_bams()
IRFinder(bams$BAM, bams$sample,
  reference_path = file.path(tempdir(), "Reference"),
  output_path = file.path(tempdir(), "IRFinder_output")
)
}
