% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CollateData.R
\name{MakeSE}
\alias{MakeSE}
\title{Constructs a NxtSE object from the collated data}
\usage{
MakeSE(collate_path, colData, RemoveOverlapping = TRUE)
}
\arguments{
\item{collate_path}{The output path given to CollateData() pointing to the
collated data}

\item{colData}{A data frame containing the sample annotation information.
Note that the first column must contain the sample names. If the names of
only a subset of samples are given, then \code{MakeSE()} will construct the SE
object based only on the samples given. Omit \code{colData} to generate an SE
object based on the whole dataset. The colData can be set later using
\code{colData()}}

\item{RemoveOverlapping}{(default = TRUE) Whether to filter out overlapping
introns of IR events belonging to minor isoforms. MakeSE will try to
identify which junctions belong to major isoforms, then select the
junctions from non-overlapping minor isoforms in an iterative approach,
until no non-overlapping introns remain. This is important
to make sure IR events are not 'double-counted'}
}
\value{
A NxtIRF SummarizedExperiment (\code{NxtSE}) object
}
\description{
MakeSE() creates a NxtSE object from the data collated
from IRFinder output using \code{\link[=CollateData]{CollateData()}}.
}
\details{
NxtSE is a specialised class (inherited from the
\linkS4class{SummarizedExperiment} class) used in NxtIRF
for downstream analysis. It contains processed data collated from IRFinder
output from multiple samples. Quantitation of included or excluded
IR and alternative splicing events, as well as relevant QC parameters
are included. See \link{NxtSE-methods} for further details.
}
\examples{
se = MakeSE(collate_path = file.path(tempdir(), "NxtIRF_output"))
}
