% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildRef.R
\name{Mappability-methods}
\alias{Mappability-methods}
\alias{GenerateMappabilityReads}
\alias{GenerateMappabilityBED}
\title{Creates a custom Mappability Exclusion BED file}
\usage{
GenerateMappabilityReads(
  reference_path,
  fasta_file,
  read_len = 70,
  read_stride = 10,
  error_pos = 35,
  verbose = TRUE
)

GenerateMappabilityBED(
  reference_path,
  aligned_bam = file.path(reference_path, "Mappability", "Aligned.out.bam"),
  threshold = 4
)
}
\arguments{
\item{reference_path}{The directory to store the reference files}

\item{fasta_file}{The path to the user-supplied genome fasta file}

\item{read_len}{The nucleotide length of the generated reads}

\item{read_stride}{The nucleotide distance between sequentially generated
reads}

\item{error_pos}{The position of the generated error within the reads}

\item{verbose}{Whether additional status messages are shown}

\item{aligned_bam}{The BAM file of reads (generated by
\code{GenerateMappabilityReads()}), aligned by the user using their
aligner of choice.}

\item{threshold}{Regions with this read depth (or below) are defined as low
mappability regions.}
}
\value{
None. \code{GenerateMappabilityReads()} writes \code{MappabilityReads.fa}
to the given \code{reference_path}. \code{GenerateMappabilityBED()} writes
its output BED file as named by \code{output_file} inside \code{reference_path},
appended by the ".txt" suffix. This BED file can be used as input
as \code{MappabilityRef} in \code{BuildReference()}
}
\description{
These function create a custom Mappability Exclusion BED file from the
supplied genome FASTA and gene annotation GTF file.
}
\details{
Creating a Mappability Exclusion BED file is a three-step process.
\cr\cr
First, using \code{GenerateMappabilityReads()},
reads are systematically generated by the given genome FASTA file
(or the \code{reference_path} containing the genome resource).\cr\cr
Second, an aligner
such as STAR (preferably the same aligner used for the subsequent RNA-seq
experiment) is required to align these reads to the source genome. Poorly
mapped regions of the genome will be reflected by regions of low coverage
depth.\cr\cr
Finally, the BAM file containing the aligned reads is analysed
using \code{GenerateMappabilityBED()}, to identify
low-mappability regions to compile the Mappability Exclusion BED file.
}
\section{Functions}{
\itemize{
\item \code{GenerateMappabilityReads}: Generates Mappability reads from a 
genome FASTA file. This function replicates the functionality of 
generateReadsError.pl in vanilla IRFinder.

\item \code{GenerateMappabilityBED}: Generate a BED file defining 
low mappability regions, using reads generated by 
\code{GenerateMappabilityReads()}, aligned to the genome.
}}

\examples{

# (1) Creates resource files and systematically generate reads based on
# the NxtIRF mock genome:

GetReferenceResource(
    reference_path = file.path(tempdir(), "Reference"),
    fasta = mock_genome(), gtf = mock_gtf()
)
GenerateMappabilityReads(
    reference_path = file.path(tempdir(), "Reference"),
)

\dontrun{
# Setting generate_mappability_reads = TRUE to GetReferenceResource() will
# run GenerateMappabilityReads after resource retrieval. The following code
is equivalent to that of above:

GetReferenceResource(
    reference_path = file.path(tempdir(), "Reference"),
    fasta = mock_genome(), gtf = mock_gtf(),
    generate_mappability_reads = TRUE
)

# (2) Align the generated reads using Rsubread:

setwd(file.path(tempdir(), "Reference"))
Rsubread::buildindex(basename = "./reference_index", 
    reference = mock_genome())
Rsubread::subjunc(
    index = "./reference_index", 
    readfile1 = "MappabilityReads.fa", 
    output_file = "MappabilityReads.bam", 
    useAnnotation = TRUE, 
    annot.ext = mock_gtf(), 
    isGTF = TRUE
)

# (3) Analyse the aligned reads for low-mappability regions:

GenerateMappabilityBED(aligned_bam = "MappabilityReads.bam",
    output_file = file.path(tempdir(), "Reference", "Mappability.bed")
)
}
}
\seealso{
\link{BuildReference}
\url{https://github.com/williamritchie/IRFinder/blob/master/bin/util/generateReadsError.pl}
}
